+++ Keyboard_Avoidance/ViewController.swift
@@ -30,7 +30,7 @@ class ViewController: UIViewController {
         .withSignUpButtonStyle
     
     lazy var content = UIStackView()
-        .vertical(spacing: 5)
+        .vertical(spacing: UI.spacing)
         .views(
             HeaderView()
                 .withFixedHeight,
@@ -61,12 +61,12 @@ class ViewController: UIViewController {
         // Hierarchy.
         view.addSubview(body)
         body.translatesAutoresizingMaskIntoConstraints = false
-        body.topAnchor.constraint(equalTo: view.topAnchor, constant: UI.spacing).isActive = true
+        body.topAnchor.constraint(equalTo: view.topAnchor, constant: UI.padding).isActive = true
         body.bottomAnchor.constraint(
             equalTo: view.keyboardLayoutGuide.topAnchor,
-            constant: -UI.spacing
+            constant: -UI.padding
         ).isActive = true
-        body.leftAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leftAnchor, constant: UI.spacing).isActive = true
-        body.rightAnchor.constraint(equalTo: view.safeAreaLayoutGuide.rightAnchor, constant: -UI.spacing).isActive = true
+        body.leftAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leftAnchor, constant: UI.padding).isActive = true
+        body.rightAnchor.constraint(equalTo: view.safeAreaLayoutGuide.rightAnchor, constant: -UI.padding).isActive = true
     }
 }